+++ src/main/java/gift/product/exception/option/ProductOptionException.java
@@ -1,11 +1,11 @@
-package gift.product.exception;
+package gift.product.exception.option;
 
 import org.springframework.http.HttpStatus;
 
-public class ProductException extends RuntimeException {
+public class ProductOptionException extends RuntimeException {
     private final HttpStatus httpStatus;
 
-    public ProductException(String message, HttpStatus httpStatus) {
+    public ProductOptionException(String message, HttpStatus httpStatus) {
         super(message);
         this.httpStatus = httpStatus;
     }
