+++ src/main/java/gift/product/exception/product/ProductErrorHandler.java
@@ -1,4 +1,4 @@
-package gift.product.exception;
+package gift.product.exception.product;
 
 import gift.common.util.ApiResponse;
 import gift.product.application.ProductController;
