+++ src/main/java/gift/product/service/ProductService.java
@@ -2,7 +2,7 @@ package gift.product.service;
 
 import gift.product.domain.Category;
 import gift.product.domain.Product;
-import gift.product.exception.ProductNotFoundException;
+import gift.product.exception.product.ProductNotFoundException;
 import gift.product.persistence.CategoryRepository;
 import gift.product.persistence.ProductRepository;
 import gift.product.service.command.ProductCommand;
