+++ src/main/java/gift/wish/service/WishService.java
@@ -3,7 +3,7 @@ package gift.wish.service;
 import gift.member.domain.Member;
 import gift.member.persistence.MemberRepository;
 import gift.product.domain.Product;
-import gift.product.exception.ProductNotFoundException;
+import gift.product.exception.product.ProductNotFoundException;
 import gift.product.persistence.ProductRepository;
 import gift.wish.domain.Wish;
 import gift.wish.exception.WishNotFoundException;
