+++ src/test/java/gift/product/domain/ProductTest.java
@@ -3,7 +3,7 @@ package gift.product.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import gift.product.exception.ProductNoConferredException;
+import gift.product.exception.product.ProductNoConferredException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
