+++ src/test/java/gift/product/service/ProductOptionServiceTest.java
@@ -9,8 +9,8 @@ import static org.mockito.BDDMockito.then;
 import gift.product.domain.Category;
 import gift.product.domain.Product;
 import gift.product.domain.ProductOption;
-import gift.product.exception.ProductOptionDuplicatedException;
-import gift.product.exception.ProductOptionNotDeletedException;
+import gift.product.exception.option.ProductOptionDuplicatedException;
+import gift.product.exception.option.ProductOptionNotDeletedException;
 import gift.product.persistence.ProductOptionRepository;
 import gift.product.persistence.ProductRepository;
 import gift.product.service.command.ProductOptionCommand;
